Update integration tests
* Make CustomUp test more stricter for nrustlightning.
  Since sometimes it starts sending request when it is not ready.
* Make DBTrieRepository null safe.
* Use common base class for LNIntegration tests.
