+++ tests/NRustLightning.Server.Tests/Extensions.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using BTCPayServer.Lightning;
 using BTCPayServer.Lightning.CLightning;
+using BTCPayServer.Lightning.Eclair;
 using BTCPayServer.Lightning.LND;
 using DockerComposeFixture;
 using DockerComposeFixture.Exceptions;
@@ -181,7 +182,8 @@ namespace NRustLightning.Server.Tests
                     CustomUpTest = o =>
                     {
                         return
-                            o.Any(x => x.Contains("Content root path: /app")) // nrustlightning is up
+                            o.Any(x => x.Contains("Listening for P2P message from any IP on port")) // nrustlightning is up
+                            && o.Any(x => x.Contains("PeerManagerProvider started")) // ditto
                             && o.Any(x => x.Contains("Server started with public key")) // lightningd is up
                             && o.Any(x => x.Contains("BTC: Node state changed: NBXplorerSynching => Ready")) // nbx is up
                             && o.Any(x => x.Contains("BTCN: Server listening on")); // lnd is up
